remove styled-components
