+++ package.json
@@ -14,8 +14,7 @@
     "aws-amplify": "^6.10.3",
     "jest-environment-jsdom": "^29.7.0",
     "react": "19.0.0",
-    "react-dom": "19.0.0",
-    "styled-components": "^6.1.13"
+    "react-dom": "19.0.0"
   },
   "devDependencies": {
     "@aws-amplify/backend": "^1.10.0",
