+++ src/pages/cv/Cv.tsx
@@ -1,34 +1,33 @@
-import { use } from 'react'
-import styled from 'styled-components'
+import { PropsWithChildren, use } from 'react'
 import * as cv from '../../assets/cv.json'
 import { images } from './logos'
 import { LanguagesContext } from '../../languages/LanguagesContext.ts'
 import { languages } from '../../languages/languages.ts'
 import { MOBILE_WIDTH } from '../../assets/constants.ts'
 
-const Container = styled.div`
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding: ${({theme}) => theme.padding};
-`
+const Container = ({ children }: PropsWithChildren) => <div>{children}</div>// styled.div`
+//     display: flex;
+//     flex-direction: column;
+//     align-items: center;
+//     padding: ${({theme}) => theme.padding};
+// `
 
-const TextContainer = styled.div`
-    max-width: ${MOBILE_WIDTH}px;
-    white-space: pre-wrap;
-`
+const TextContainer = ({ children }: PropsWithChildren) => <div>{children}</div>// styled.div`
+//     max-width: ${MOBILE_WIDTH}px;
+//     white-space: pre-wrap;
+// `
 
-const ImgContainer = styled.div`
-    display: flex;
-    flex-direction: row;
-    overflow-x: scroll;
-    max-width: 100vw;
-`
+const ImgContainer = ({ children }: PropsWithChildren) => <div>{children}</div>// styled.div`
+//     display: flex;
+//     flex-direction: row;
+//     overflow-x: scroll;
+//     max-width: 100vw;
+// `
 
-const StyledImg = styled.img`
-    height: 50px;
-    padding: ${({theme}) => theme.padding};
-`
+const StyledImg = ({ children, ...props }: PropsWithChildren) => <img {...props}>{children}</img>// styled.img`
+//     height: 50px;
+//     padding: ${({theme}) => theme.padding};
+// `
 
 type CvRow = {
     heading: string
